AttachmentStatus missing "available" (#273)
I mean this is clearly not in the API docs either, so there's that, too.

The latest API version I can find is [API Version 2015-10-01][APIRef],
and it clearly fails to mention "available" as a state. On the
management console, there are two states, too.

1. State: attached
2. Attachment state: available

I'm just going to guess that this means it's attached to the VPC, and
it's available for routing traffic, but that's definitely a guess.

I've also just tested that this change fixes this error.

Thanks!

[APIRef]: http://docs.aws.amazon.com/AWSEC2/latest/APIReference/API_InternetGatewayAttachment.html "Latest API for InternetGatewayAttachment"