Strip 'Expect' header when presigning URLs
Fixes #363.

This is more or less the _simplest_ thing that can be done to make
pre-signed PUT urls work, and it exposes the header transformation
at the lowest level, so that it's possible to build some other
behavior on top of that. However, this is neither as targeted nor
as flexible as the solutions proposed in the bug.