Add support for extra headers as needed by Glacier
Today, the code generated to interact with AWS Glacier does not include the
required header `x-amz-glacier-version: 2012-06-01`, so calls to Glacier fail
with the message `Required parameter missing: API version`. This commit fixes
this by adding support for arbitrary extra headers in the code generator, and
updating the annex for Glacier to include this header.

Fixes #402