Drop some MonadThrow and MonadCatch constraints
This is arguable, and simply closing is acceptable. My argument for
making this change is that, when `MonadIO` is already in the
constraints, adding in either `MonadThrow` or `MonadCatch` simply adds
ambiguity as to where exceptions may end up being thrown. Reducing to
just `MonadIO` makes it clear that all exceptions will be thrown as
runtime exceptions.

There may be other places in the library this approach applies, but this
is where I noticed it.