Remove duplicate files on case-insensitive filesystem.
When checking out this project on MacOS, the following warning
appears:
> warning: the following paths have collided (e.g. case-sensitive paths
> on a case-insensitive filesystem) and only one from the same
> colliding group is in the working tree:
>
>   'amazonka-iot/fixture/CreateCertificateFromCSR.yaml'
>   'amazonka-iot/fixture/CreateCertificateFromCsr.yaml'
>   'amazonka-iot/fixture/CreateCertificateFromCSRResponse.proto'
>   'amazonka-iot/fixture/CreateCertificateFromCsrResponse.proto'

In addition, `nix-prefetch-git` returns a different SHA256 hash
on Linux, or MacOS, because the state of the checked out directory
is different (two files are missing on MacOS because of the
collision).

This commit keeps the files with the capitalization from tests.