added: load region data from aws file when use AWS_PROFILE
I was having trouble with the following error when I was programming.

```
message    = Just The bucket you are attempting to access must be addressed using the specified endpoint. Please send all future requests to this endpoint.
```

The necessary permissions are given to the user.
For some reason,
it is working in the production environment (ECS).

After investigating,
I found out that this problem occurs only when I pass the authentication data using the environment variable `AWS_PROFILE`.

So I read amazonka's code and found out that `Region` is not set via environment variable.
So I referred to
[Configuration and credential file settings - AWS Command Line Interface](https://docs.aws.amazon.com/cli/latest/userguide/cli-configure -files.html)
and configured it so that the setting is done automatically.

Note: This commit fails to build in my environment,
but that's a problem with the text-icu package.
[Build breaks with icu 68 - Issue #49 - haskell/text-icu](https://github.com/haskell/text-icu/issues/49)
It worked when I configured it to load the modified package.