amazonka-rds: Support DestinationRegion pseudo-parameter
Cross-region replication on RDS is performed by making a request
against the _destination_ region. In that request, you need to pass a
`PreSignedUrl` for the _source_ region, with the `DestionationRegion`
parameter set to the _destination_ region.

There are five such requests at the moment, specified by
https://github.com/boto/botocore/blob/77359646878e261eaa9debad888fc8f3b20832ef/botocore/data/rds/2014-10-31/service-2.sdk-extras.json

These specify a `SourceRegion` parameter because botocore rearranges
the rds requests internally, stripping `SourceRegion` from the request
itself and injecting a `DestinationRegion` parameter into the
`PreSignedUrl` field:

https://github.com/boto/botocore/blob/ec0eb680d6e6e753c7d7ddf0242b8ae1c3b38db5/botocore/handlers.py#L1080-L1089

Fixes #378