Stop chunked bodies causing empty ContentEncoding metadata in S3
https://docs.aws.amazon.com/AmazonS3/latest/API/sigv4-streaming.html
says that chunked requests should set `Content-Encoding: aws-chunked`.
S3 strips any occurrence of `aws-chunked` from that header, and stores
the remainder as the `ContentEncoding` metadata on the object. It
isn't very smart about this: if `aws-chunked` is the only encoding
listed, S3 will store the empty string as the encoding, which breaks
some HTTP clients and CDNs.

Multiple AWS SDKs appear to have hit this issue, and the developers of
`fog` contacted AWS support. Their suggestion was to omit the
`Content-Encoding:` header entirely because S3 is capable of figuring
it out: https://github.com/fog/fog-aws/pull/147