Control VHost url rewriting with a `Service` field
Originally "Non-working attempt to fix the test failure in PR #797"

This adds a new field to `Service` which controls if requests to the
service should have their URLs rewritten using `s3vhost`. This is
particularly relevant for mocks of S3 which don't support having the
bucket name in the hostname (e.g. because they are bound to `localhost`,
which isn't a domain and can't have a subdomain.)

This field is defaulted to `True` to preserve the current behaviour for
all services.

Co-authored-by: Isaac van Bakel <isaac@supercede.com>