Use the same `Service` for all sub-requests
This implements a change suggested by this comment:
https://github.com/brendanhay/amazonka/pull/797#issuecomment-1181531912
as part of a fix to the following issue:
https://github.com/brendanhay/amazonka/issues/760

This modifies the code generation for AWS endpoints so that requests
which make other requests all share the same (passed-in) `Service`. This
is done by making `request` take an extra `Service` argument. This also
adds a new parent class of `AWSRequest`, called `AWSService`, which
defines the default service used for each request (this being
`defaultService`.)

These are some breaking changes, but they are required to allow modified
`Service`s to be used in complex requests. This is necessary to allow
VHost rewriting to be turned off for local mocking and other use cases
at the `Service` level.

Co-authored-by: Bas van Dijk <v.dijk.bas@gmail.com>