gen: Compute loop-breaker edges using rewrittenService
Some named shapes are not written out as modules, because they denote
a simple list or map of other shapes. By checking for loops on the
result of rewriteService, we correctly `import {-# SOURCE #-}` where
we otherwise wouldn't.

Also do not worry about cutting loops from a type to itself, as while
they don't cause import problems, the presence of uncompiled
`.hs-boot` files breaks the Bazel build.