amazonka-core: use %Z when parsing timezones to support non-GMT
It seems that AWS (maybe just S3?) can sometimes return RFC5322-format
dates with an `obs-zone` other than "GMT", so widen our parser to
accept them.

A `%Z` in the format string causes `formatTime` to write the time zone
as `UTC`, which is not a `zone` under RFC 822 nor an `obs-zone` in
RFC 5322. We therefore split the `TimeFormat` class to distinguish
between parsing and printing formats, so that we continue to emit the
`obs-zone` of `GMT`.