amazonka(hooks): Add monomorphic hook-modifying functions per field
The polymorphic hooks in 2.0 are easy to misuse. In amazonka-2.0, to
apply a response hook requires getting the argument type exactly
right:

```
addHookFor_
  @(Request ListBuckets, ClientResponse ListBucketsResponse)
  (...)
```

If you got this type even slightly wrong, your code would compile and
your hook would silently never fire.

Therefore: deprecate the polymorphic hook-editing
functions (`addHook`, `addHookFor`, etc.) and provide functions to
modify the hook at a particular field.

Where the hook is polymorphic in the request type, provide two "add"
functions: one that runs the hook against values of every type, and an
"add for" function that runs the hook against a particular AWS request
type.

Also provide a "remove for" function to block hooks from firing for a
particular AWS request type. `noHook` and `noHook_` can suppress a
hook for every request type, so we don't need to provide versions of
that for each hook field.